requirements.txt
#requests==2.27.1
#beautifulsoup4==4.11.1
selenium==4.4.0
webdriver-manager==3.8.3
python-dotenv==1.0.0